+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - W3C Common PSSH Box UUID
 
+### Added
+
+- Partial parsing of HEVC VPS
+
 ## [0.41.0] - 2024-01-12
 
 ### Added